\name{FIBS-package}
\alias{FIBS-package}
\alias{FIBS}
\docType{package}
\title{FIBS package
\packageTitle{FIBS}
}
\description{
\packageDescription{FIBS}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{FIBS}
\packageIndices{FIBS}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{FIBS}

Maintainer: \packageMaintainer{FIBS}
}
\references{
~~ Literature or other references for background information ~~
}

\keyword{ package~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~ }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
