\name{model3}
\alias{model3}
\docType{data}
\title{Model 3: High versus low disturbance (Highgrove and Laxton)
}
\description{ The values of the two functional ecological attributes of the quadrats of modern arable fields from Highgrove and Laxton, UK.

}

\format{
  A data frame with 36 observations on the following 4 variables.
  \describe{
    \item{\code{Survey location}}{Location of the modern arable fields}
    \item{\code{REGIME}}{Grouping variable for type of farming regime (1 or 2)}
    \item{\code{FLOWPER}}{Length of flowering period}
    \item{\code{VEGPROP}}{Vegetative propagation}

  }
}

\source{Stroud et al in prep
}
\references{
Bogaard, A., Hodgson, J., Kropp, C., McKerracher, M. and Stroud, E. (in press 2022). ‘Lessons from Laxton, Highgrove and Lorsch: Building arable weed-based models for the investigation of early medieval agriculture in England’, in M. McKerracher and H. Hamerow (eds) \emph{New Perspectives on the Medieval ‘Agricultural Revolution’: Crop, Stock and Furrow} (Liverpool: Liverpool University Press).

}

\keyword{datasets}
