\name{Weedmodels}
\alias{weedmodels}
\alias{FIBS}
\alias{weedmodels}
\docType{package}
\title{Weedmodels
\packageTitle{Weedmodels}
}
\description{
\packageDescription{Weedmodels}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Weedmodels}
\packageIndices{Weedmodels}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{Weedmodels}

Maintainer: \packageMaintainer{Weedmodels}
}
\references{
~~ Literature or other references for background information ~~
}

\keyword{ package~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~ }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
